tools/importer-rest-api-specs: returning all models applicable for the schema when building models
This/the previous commit updates the Build function to return the complete list of models used in
the Schema for a given Resource such that we can write unit tests to validate it.

This also updates the model name for nested models to include the prefix of the resource name, this
is done to avoid conflicts when the same model is used in different resources within the same package.
Whilst this does technically mean we have more models in the short-term, in practice this is only an
issue in larger packages (e.g. Compute/Network) - and is something we can refactor/improve upon in the
future.